Move v3 components into v3 folder
for consistency with other existing components
