+++ apps/desktop/src/components/v3/HeaderControlSection.svelte
