+++ apps/desktop/src/components/v3/StackHeader.svelte
