+++ apps/desktop/src/components/v3/StackTab.svelte
