+++ apps/desktop/src/components/v3/StackTabPreview.svelte
